requirements.txt
MoviePy
imageio
imageio-ffmpeg
numpy
pillow
opencv-python
